chore: add pylint
