+++ pyproject.toml
@@ -52,6 +52,7 @@ flake8-bugbear = "==21.3.2"
 flake8-mutable = "1.2.0"
 pytest-github-actions-annotate-failures = "==0.1.1"
 flake8-pytest-style = "1.4.0"
+pylint = "^2.7.4"
 
 [tool.isort]
 default_section = 'THIRDPARTY'
