+++ setup.cfg
@@ -1,0 +1,2 @@
+[pylint.MASTER]
+max-line-length=88
